nllb_translate_server_customtext/nllb_translate_server_dual_mode/requirements_gpu.txt
transformers>=4.37
opencc-python-reimplemented
flask
flask-cors
requests
numpy<2.0